Injection doesn't like anonymous classes
